Dash test

+++ db/migrate/20140810030123_add_attributes_to_city.rb
@@ -1,8 +1,8 @@
 class AddAttributesToCity < ActiveRecord::Migration
   def change
     add_column :cities, :bigger_thing, :string
-    add_column :cities, :lon, :integer
-    add_column :cities, :lat, :integer
+    add_column :cities, :lon, :float
+    add_column :cities, :lat, :float
     add_column :cities, :country, :string
   end
 end
